+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.26.6 - 2018-09-19
+## Fixed
+- Added `Cache-Control` headers to files served by `Dash.serve_component_suites`. [#387](https://github.com/plotly/dash/pull/387)
+- Added time modified query string to collected components suites resources.
+- Added `InvalidResourceError`. [#393](https://github.com/plotly/dash/pull/393)
+- Added a flask errorhandler to catch `InvalidResourceError` from `serve_component_suites` and return a 404.
+
+## 0.26.5 - 2018-09-10
+## Fixed
+- Fix `get_asset_url` with a different `assets_url_path`. [#374](https://github.com/plotly/dash/pull/374)
+
 ## 0.26.4 - 2018-08-28
 ## Fixed
 - Set `url_base_pathname` to `None` in `Dash.__init__`. Fix [#364](https://github.com/plotly/dash/issues/364)